fix(router): move non-route game files under app/_game to avoid Expo Router route warnings (Epic #14, closes #71) (#78)
