+++ app/_game/rules.ts
